% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig.R
\name{nmc_eig}
\alias{nmc_eig}
\title{Nested Monte Carlo Expected Information Gain}
\usage{
nmc_eig(design, model, N = 1000, M = 1000, cores = 1, update_model = TRUE, ...)
}
\arguments{
\item{design}{a dataframe of independent variable values}

\item{model}{a fitted brms object}

\item{N}{int, how many outer loop Monte Carlo samples}

\item{M}{int, how many inner loop Monte Carlo samples}

\item{cores}{int, how many cores and chains to run MCMC with}

\item{update_model}{bool, update model with new samples}

\item{...}{optional parameters passed to \code{update_model}}
}
\value{
expected information gain of design
}
\description{
Esimate the expected information gain (EIG) for a given design (data frame with potential independent variables to collect)
}
\details{
#' Uses nested Monte Carlo estimation of the EIG.
See equation 4 in https://arxiv.org/pdf/1903.05480.pdf or
equation 9 & 10 in https://arxiv.org/pdf/1108.4146.pdf
}
